add ftmMain
